Updates, Added Emperor's Children and Raven Guard, added Fortifications copy
Both Legions added to the Crusade Army list.

Horus Heresy .gst altered to allow Lords of War in those FOC's that
allow it.  This is to allow Primarchs to interact as they should with
the lists - especially re: Rites of War.

Djinn-skein upgrade moved for Magos Prime.

Fortifications copied from wh40k repo.  Edited for compatibility.

Closes #2 and closes #7.