http://www.forgeworld.co.uk/resources/fw_site/fw_pdfs/Horus_Heresy/Betrayal_FAQ_Errata_V2.pdf
Q: In the case of the ‘Pride of the Legion’ Rite of War, Veteran and
Terminator squads are Troops choices in the force; could you clarify the
exact squads which fall into these categories? A: In this case
specifically this means Veteran Tactical squads and Legion Terminator
squads only, so Justaerin and Deathshroud Terminators are not included
in this. Any Legion-specific units in future Horus Heresy books which do
qualify for this Rite of War will have this fact explicitly stated in
their unit description.