Add files via upload
Little bit of a play to add in some of the missing options to the Baneblade types of tanks. Taken the Reference from HH : Crusade Imperialus (Redbook for Solar Auxila) as they have the super heavy command tank option and armoured ceremite options, due to being 30k versions. Also reduced Banehammer cost as a mistake had been made and it was showing as the same as Baneblade.
Shadowsword / Stormblade given Smokelaunchers and Spotlights