Add files via upload
Fixed some issue on Ebon Keshig
Legion Astartes
Sicaran Omega: points reduced, and now able to be taken in squads of 2
Sicaran Punisher: points increased and Rolling Thunder rule changed removing the part of the rule that only allowed it not fire next turn if it used the rending shots, and now able to be taken in squads of 2
Outriders: Fixed Power Weapon issue and 2 instances of twin linked bolters in entry, and standard war gear requiring selection
Fixed various issues involving restrictions on Jump Packs having not been done correctly. This was not disabling the option to take them on various consuls and Iron Father / Warsmith selections.