#1554 #1555 and a couple of QoL fixes for Spacewolves.
#1555 Fixed issue of having a RoW for main army and Allied detachment. This is allowed, it was just about 50/50 split the code had put max 1 Roster in some and max 1 Force in others. They should have all been Force.

#1554 Speaker of the Dead / Deathsworn issues. All MoL were being given access to Deathsworn as Command squads. It should only have been Speakers of the Dead and ONLY if they are the armies Warlord. This is fixed.
Other Quality of Life changes are:
No longer needing to select every grey slayer to have a bolt pistol. Shuffled a few boxes around and made it clear that they can exchange the pistol for a different pistol or combat weapon for a different combat weapon.
Varagyr Termys also have a default selection added to their frost weapon to be frost axes. You had to choose one anyway, so now you don't have to manually click them all to give them axes, but can still give them any of their normal options with the same amount of clicks as it used to be.