Eldar (Fanmade)
This adds the Eldar Fanmade rules from the AUS30K guys.
I've tested it a fair bit and was all working great a month or 2 ago. Seems to still be working.

I was hoping to put this up at the same time as the Necron one. But I've not been able to work on it in the last 2 weeks, and a mystery problem seems to have occurred with BS that i will need to find out why it no longer wants to work with BS.