Blackshields Fix
Fixed issue that Blackshield stuff was showing for none Blackshield legions.
For some reason BS has taken to ignoring hidden functions that are outside of some groups / entries when they are cross referencing.
As such i have put Blackshields in the "Legions" group. This group is now as such renamed "Legions & Blackshields". This seems to have fixed the issue from what i can tell.
Also small play with Vets to remove need for link within a link within a link.