1617, 1618, 1619
Additional fixes to fixes...

Ok. Few bugs in the last one it seems, as pointed out by @capitaladot

So Sabre entry is now listed as a squadron
Neutron Blaster also has S9 (for some reason when i do a whole heap of changes in 1 go, every so often it misses some of the numbers i type in to profiles)
Combi Weapons for Seekers is laid out better now. Like how they are in vets.
Combi Weapons for Mor Deythan also have had the same treatment. Rather than selecting "Combi Weapon" a dozen times then choosing which one, you just pick which ones you want.
The cost of Graviton Guns for Support squads was also adding an additional 15pts on to the cost. This has been sorted. It seems when it was added in, the person didn't look at the other entries in the same list, and add a decrement cost to them to offset the standard points value added to the entry it linked to.