Mechy Fixes
Went to a few events lately and talking with people.
Some people who only used particular units when Cybernetica but not any other were double buffing their Initiative bonus for the cybernetic as the bonus was already being applied. So i have now instead of set increment on the units. It now sets it to the new number, but also has a * by the stat. This also cross links with the wording for the Cybernetica buff, with something like "*Included in profile on BS" added to the end.
In addition also fixed the Machinator Arrays and Abayants to also have the same effect, and extra bit on description. This had been confusing people a bit as the mech book has Draykavac & Decima having the bonus included on page, but Scoria hadn't had the bonus included.
Also in doing this i noticed that the full Machinator Array rules had not been included in all 3 of those characters (so missing weapon profiles and stuff).