Add files via upload
Maddening Swarm fix for Daemons. Relinked it to fix an issue where they weren't getting their additional emanation. (Though still can't get the mandatory ones to be mandatory 

#1656 also fixed. Was an issue for some reason the Veteran marines had been clicked to be a collective, no idea why or when this happened.
Also found that it wasn't enforcing the min level of bolt pistol or exchanges for them, due to ticked "and all child forces" on the min requirement.