#1665
Had to reduce the cost of the Consul upgrade by 35pts as the Consul has Cata armour by default and the way the armour gets its value seemed unable to then be modified by the consul being taken. So the reduction has had to be factored in on the consul itself to balance it out.