Small Fixes for Dreads
Fixed the Dread / Comtemptor / Cortus. Previously by swapping your ranged weapon for a melee one, it was giving you the bonus attack, even if you then swapped the melee one you started with for a ranged one.
Also relinked Meltagun on a few things, as it was once again doubled up for no reason.