Updated a whole bunch LoW restrictions
A whole bunch of options were listed as not being hidden for LoW when Delegatus, Day of Revelation or Ravenwing Protocol was selected.

Issue: Can't hide the knights. I might have to adjust how the GST file contains them. So that they are once again shared entries, but not as shared root entries. This was changed some time ago.