Bunch of changes to fix upcoming issue with new BS
Moved Knights root entries in to individual lists, rather than GST files. This fixed an issue for Legion Astartes that some RoW didn't allow knights, but were still able to take them. Also same with Knight lists as they are only allowed to take Titans as LoW. But was done to fix the issue of in the current build Beta build of BS.