Mournival Units
Need to bear with me on this one. This is another Fanmade one from the AUS 30K / Mournival guys.
These files will cause no conflicts with the other files, as they have been kept sepparate.
I have added in all the units from the Mournival 3.0 digital copy for Marines. I will continue to add in the other units, and then finally get on to the revised units from the FAQ. Though i am unsure how i will be doing this, as there are 2 options. The short way is just make a set of "Revised selections" as a category on this file. that mean that you can select an option such as
Iron Warriors: Iron Havocs unit cost reduction. And that will be charged at -25pts

In the case of terminators getting the option for Volkite Chargers, I can have an option of
Exchange Combi-bolter for Volkite Charger. That will be charged at 2pts a time.

All these options will have the weaponry linked within them for profiles and rules, so they will still go on the summery at the bottom.

Anyway hope you enjoy.