Few Fixes, lots of slimming, and a better Mournival
As title.
Major overhaul is the Mournival stuff. This is a BIG change as you no longer need to make a separate listing for your Mournival stuff. The catalogue is now linked to the LA CA one, so it can share info. This means all the entries for LA are going to be on 1 list. THIS also means given a bit of time, i will be able to add in all the RoW somehow, and you will be able to make working lists.