Quick Fixes and more Custom Char stuff
Quick fix for Blackshield Marauder units (they weren't tagged as compulsory
Added some rules lines to Cadre weapons for World Eaters that were missing such as Two-Handed or Rending...
Added some rules lines to Raven's Talons and Pair of Ravens Talons as they were missing all the special rules for them.
Fixed restrictions on Cameleoline for Ravanguard for Centurions. Previously it was only preventing you from taking it when you had terminator armour. It should have also prevented it when using jump packs, bikes or jetbikes.

Mournival Custom Char stuff. Added in all the remaining Legion Equipment. So this just leaves the biggest and hardest chunk left to go, and that is the Specialisation stuff. This will no doubt take a few weeks to implement fully..