Salamanders Praetor Mastercrafted
Option for Artificer Weapons was missing from Legion Options. However "Praetor May Take:" changed the cost of Mastercrafting down to 5pts. This however is not the same thing, as a Salamander Praetor could take both Artificer Weapons (5pts) & Master-Craft a Weapon (15pts) at the same time if they wished.
The likely chance is that this was not taken often due to the dragon shield option being available. But there might be an odd situation that someone may have taken a Thunderhammer and Paragon Blade (or lightning claw) or something like that to have a 2X Strength weapon often instant deathing or very high toughness hitter and a general troop killer in the other hand.