Change to Option for Playtest Rules Errata 1.0
Rather fed up with people keep saying "The rules are wrong" or "You haven't put in the FAQ Rules"...
So "Playtest Rules" is now named "Playtest Rules Errata 1.0 (From FAQ 1.1 Feb/2019)" and the default for it is selected to On. As I believe near universally people play with them in game, even though they are still PLAYTEST rules, rather than officially approved.