Many changes and #1736 Alliess issue
Fix for Alliess issue with using Ordo Reductor allies to another force.

Also had a mess around to make list building easier and making it it easier to edit Mech CAT in future changes. Though this will mess up some lists unfortunately.

Basically there were several double entries for thing like Adsec Thralls, Peltast, Hoplites, Scylax Guardians, Magos Prime and Dominus, Casstellax,& Thallax.
A feature introduced some time ago in editor was to add catagories. This allows me to select when something gets Compulsory or not based on an option selected. So now they are fixed.
This should make it easier to make a roster if you forget to select the Legio before selecting units. As previously anything you selected could then be told that that option had to be hidden as the other version was now required.