issue #1743 update
Mirrored change to add Compulsory Elites category added for Garrison Raiders and ZM Attacker forces. Only 1 Elite that I know of always counts as Support Squad (Ebon Keshig) So all the rest have it added to them. Tried to tie the adding of that category to those 2 force org charts, but it wouldn't allow (or i couldn't work out why it wouldn't).