Compulsory (Went off the rails!)
Played around with the compulsory entries. Tact, Assault and Breachers now only have 1 root entry. Each of which will be tagged as (Compulsory) and have the catalogue entry of compulsory if certain RoW are not selected. This data was taken from the original checklist for hidden/show files. Though may need a review to make sure all the RoW are actually applying it correctly.
Found that it seemed to have been bugged at some point during the Dark Angels branch update. Unsure when or why, but at least 1 of the "If Equal to X" conditions was no longer working, but unsure which. They are now working for all 3 of them.