Various Fixes Part 3
Various units Dedicated Transport options corrected. This was mainly to do with Steel Fist and Ironwing interactions. Previously it was just coded to show if Steel Wing was selected, without requiring the component of the Scion of Ironwing being present. This was corrected.
Contekar fixes so the claw now costs 15pts. Also available as a Reward of Treason.
Deathwing Companion Melta Bomb cost fix.