Command Units (Sorry about another change)
Ok, so little / big change here to make the Bodyguard / Command Squad / Retinue thing work better.
Termy Command Squads can only be taken if the Praetor or other MOTL warlord has Termy Armour and is also the army warlord, or is a Primarch (as per FAQ). Specialist versions such as Justaerin, Firedrakes, Deathshroud and Deathwing Terminator Companions also have this rule.
Removed ability to select units from the Bodyguard section if Primarch is selected, with the exception of Command Squad and Termy Command Squad. The cost of others are currently counted as part of the cost of 25% LoW limitations. Until they also get an FAQ or reprint to say the same as Command Squads did).

Also added Infantry to Indomitus Termy armour