Daemonic Shenanigans
So I've just added all the descriptions to all the Dominions.
In addition I have added Daemonic Ambiguity Some think that the rule for Creeping, Crimson, and Lurid in regards to their special Emanation means that all units containing more than 1 model Must take the Emanation. Others believe that it is Optional. So their is now a sub group menu to selection Must or Optional.