Updates
Removed the (Salamanders) versions of heavy flamers, and relinked all to the heavy flamers in the Legion file, which now works correctly to give the +1S it should have. Also made any hidden that should be due to change. Also found a few additional ones that hadn't been changed.
Fixed a few LoW entries to have their weapon profiles.
Fixed Justarian weapon profiles
Fixed a couple of tanks that start with Heavy Bolters (such as Sicarian variants) with heavy bolters to be able to take the flamer for sallys.
Fixed Firedrakes to be able to still take the heavy flamer if taken by Alpha Legion.
Added Twin-Linked Heavy Stubber to GST for some entries that required it.