Ironwing Protocol fix
Ironwing Protocol kind of fixed.
Still no way to enforce the min number of transport capacity to mean all units MUST start in transports, due to things like Spartan being able to hold a squad + apoth + characters...
However, fixed the need 1 tank per 2 units thing by adding in a category to all units at the root entry "Ironwing Protocol Min 50% Tanks Requirement" Though I can probably have a more snappy name. Also this might be a useful thing to have added to all units. So instead of using "Unit" from the Unit/Upgrade/Model selection, it would be a true count of all units, including dedicated transports, which if taken as a sub entry of an existing unit, do not seem to flag as a unit themselves in cases.
HOWEVER this has meant that in the case of the Terrax, Spartan, Stormeagle, Rhino, Proteus, Phobos have all had to be copied to make a unique dedicated transport version, and a non-dedicated transport version. Which could cause some additional work when making amendments.