Test rosters to check changes worked
Test Rosters - Descriptions are in the Custom name of the list, but also listed below

BLACKSHIELDS
A Blackshield detachment may never have more units with the vehicle type than it has units with the infantry type.
3 Infantry units are in the army (Reaver Lord, Mauauder, Breacher) The dedicated transports do not count
The Sabre, Arquitors, Leviathans and Preditors are all Vehicle Units
Removing any one of the Vehicle units will resolve the issue,
OR
Adding any other Infantry unit.

BLOOD ANGELS
With the exception of Dedicated Transports, a Blood Angels Detachment may not have more units with the Vehicle type than it has units with the Legiones Astartes (Blood Angels) special rule.
3 Legion Astartes units are in the army (Preator, Breacher, Tact) The dedicated transports do not count
The Sabre, Achiles, Arquitors or Leviathans are all Vehicle Units
Removing any one of these units will resolve the issue,
OR
Adding any other Legion Astartes unit.

Hammer of Olympia
Must have a min of 1 Warsmith or Siege Breaker. This can be resolved by:
1) Giving the Centurion Siege Breaker
2) Giving the Praetor Warsmith
3) Taking Kyr Vhalen as he is a Warsmith instead of the Praetor (This was previously an issue)

IRON HANDS
Iron Hands detachment may not have more units with the Legiones Astartes (Iron Hands) special rule in total (including Independent Characters) with the Jump Infantry, Bike or Jetbike types than it does with the Infantry type.
The Infantry units in this army are Praetor, Breacher, Tactical (3 units)
The Jump, Bike, Jetbike units are Assault, Hussar, Jetbike, Outrider.
NOTE: The Assault squad counts as both Infantry for some things as well as Jump. So has a 2 modifier to the number of infantry units, as it counts as infantry itself.
To resolve the error, add 1 Infantry unit, but not a Jump Infantry unit.

IRONWING PROTOCOL
At least half the units in this RoW must vehicles with the tank type.
This list has the new dedicated transports count at part of the unit turned on. So this army has 3 units total (Praetor, Breacher, Tact), as the dedicated transports do not count as their own units, but as part of the unit confer the "Tank" type to the unit they are attached to.
If you change the list to dedicated transports count as their own unit, it will cause an issue as you will then have 5 units (Praetor, Breacher, Tact, Landraider, Rhino) of which the later 2 are all units consisting of tanks. Which is less than the 50% required.

Maru Skara
Min of 2 HQ, of which 1 needs to be a champion. Was previously giving an issue of flagging the champion issue 3 times. Only removing 1 if you had any champions I the army.
Resolution was to add a tickbox inside of Champion with a min/max of 1. This also helped the Mournival stuff.

RAVENGUARD
A Raven Guard detachment may not have more units in total with the Vehicle (Tank) type than it does with the Legiones Astartes (Raven Guard) special rule.
3 Legion Astartes units are in the army (Preator, Breacher, Tact)
The Sabre, Achiles, Arquitors & Predators are all Vehicle Units
Removing any one of these units will resolve the issue,
OR
Adding any other Legion Astartes unit.
If you change that Dedicated Transports count as their own units, it will then require the removal of 3 Tanks Units OR addition of 3 Legion Astartes units

STEEL FIST
At least half the units in this RoW must consist entirely of vehicles with the tank type.
This list has the new dedicated transports count at part of the unit turned on. So this army has 4 units total (Praetor, Breacher, Tact, Preds), as the dedicated transports do not count as their own units. So also the dedicated transports are not a unit entirely consisting of vehicles with the tank type as they have infantry in there.
To remove the error on list you need to add 1 more unit which  is a tank such as unit of predators or sabre. OR if you switch to dedicated transports count as their own unit, it will resolve the issue as you will then have 6 units (Praetor, Breacher, Tact, Landraider, Rhino, Predators) of which the later 3 are all units entirely consisting of tanks.