More test rosters added
Ultramarine Logos
Issue with force org asking for masters of signal multiple times. If one is added to this roster or a Damocles rhino it is resolved.

SOH Black Reaving and Justarian.
Issue with force org asking for masters of signal multiple times. If one is added to this roster is resolved.
Also fix for Justarian min/max weapon requirements.