Ahrimans Command Squad #1883
Ahrimans Command Squad #1883 fixed. They were lacking the ability to take the ML2 Brotherhood of Psykers when he was warlord.
Fixed by adding an additional selection entry to Ahriman then cross linking it to the original "ancestor of" options from the command squad, which were then changed to Force and that selection entry.