Chogorian fixes & #1884
If Chogorian Brotherhood RoW is selected: Added outflank to all infantry units in force org . Also individually added Hit and Run to all Infantry units with conditions to hide it if any salvo, heavy or ordnance weapon is in the squad.

Also fixes from #1884 which highlighted that "Shattered Legion" entry link, was mistakenly added to the blood angels and whitescars units.

If Chogorian Brotherhood RoW is selected: Added outflank to all infantry units in force org . Also individually added Hit and Run to all Infantry units with conditions to hide it if any salvo, heavy or ordnance weapon is in the squad.

Chogorian Brotherhood - Outflank / Hit and Run test roster included in upload.
Removing the RoW or adding a heavy or salvo weapon to any of these units will remove the "Hit and Run" rule. Or giving a unit like a Centurion or Apothicary a Bike or Jetbike will also remove it as well as the Outflank rule.