HH9 Crusade Errata Updates #1894 #1895
HH9 Crusade Errata Updates #1894 #1895

Remove Land Raider Proteus option for Deathwing Companion Terminators, replace with Spartan transport option. (Left the Landraiders in, but only if Steel Fist RoW is in, as it would override their DT preferences)

Remove Hatred and add Preferred Enemy (characters) to Firewing Enigmatus Cabal.

Change Dreadwing Interemptors transport parameters to "10 models or less" rather than exactly 10

Add "This counts as a sword for the purposes of the Mastery of The Blade" to Corswain's The Blade's item description.

Remove the Inbuilt option for the Lion to take Deathwing Companions. Instead give him the same check that allows for Deathwing Companions and Deathwing Companion Terminators to be taken in the "Command Squads" section.

Updated rules and restrictions for the Storm of war to reflect changes.