Lots of little legion fixes (reupload)
#1920 Thousand sons psyker level can be bought by psykers
Excluded consul Nullificator, Esoterist and Librarian from selecting option

#1921 Space marine crew missing in sub orbital strikewing
Added the option to Sub Orbital Strike wing, however it will not apply the buff to the statline cos of the way it is formed. May look for alternative solution later such as separating the sub orbital strike win into its component parts.

#1922 Point cost of battle servitor controls missing in sub orbital strikewing 
Added repeating cost based on the number of each of the types of planes in the wing. Previously this was linked to the "model" count, however when the "units" change was put in, this inadvertently broke the math.

#1923 Legion exoterist missing psyk-out grenades
Added Exoterist to the flag list for unhiding the grenades and added a cost of 5pts for them.

#1925 Veterans in a chosen duty rite of war do not count as minimum compulsory
Fixed a larger issue that seemed to be affecting this one. It seemed in all cases it wasn't applying the veteran unit as compulsory choice.
 master