Vigil Stormshield Ambiguity
Finally added the Ambiguity for Command Squads and Nullificators for Imperial Fists. This was previously overlooked for a long time due to the number of bug reports endlessly coming in about it. Now it is up to you to make the judgement and select the option of if they can or can't, as some local groups use it one way or another. Though some will trust Battlescribe knows the rules in these cases when possibly not officially clear.