Pride of the Legion
Added Nullificator as troops to all legions if Primus Nullificator is taken, and removes the 0-1 restriction.

Added new Selection entry “Pride of the Legion Compulsory Troops” with a hidden tick and link to new category of same name in Legions Astartes Lib (because it doesn’t work to check max in force 2 otherwise…(this may need modification if other FoC come out)), which shows if PotL is in force, and max in parent is set to 1. This is added to Vets, Cata & Tart squads.
All 3 units also get Add Category Line and Compulsory Troops if that option is selected.
Additionally Indomitus Termy have Support squad removed, compulsory troops and line added if PotL.
Made Crimson Paladins Troops for BA (This will need replication in AL, and all other legions need this for their terminator armoured elite choices)

Had to move some category modifiers up one level in Centurion and Praetor, as they weren’t recognised for restrictions otherwise.
Note: Shamshir needs adding to Centurion, and also then the modifiers for Infantry and Cavalry will needs amending to include it.