Liber Imperium FAQ updates and a little bit more
Assassins updated (though noticed that one of the FAQs that updated a rule, the rule isn't actually referenced anywhere... (Page 136 The Shadow of Death: Change ‘Primarch Sub-type’ to ‘Primarch Unit Type’.) The rule exists and seems to be for assassins, but none of the assassins have it...
Custodes (placeholders) updated (Achilles Dread and Tarsus Buckler)
SoS some minor updates to add units. Up to Troops now.
Solar FAQ updates. A load of Legecy units still to do.