LA and WE patch for Infantry / Non-Infantry
There were some issues with the infantry / non infantry settings.

It was applying the restrictions to every army 
It also also keeping units tagged as Infantry even if they had dedicated transports.
It was also counting in some instances multiple of something in a unit as multiple instances in a unit of being non-infantry in some dreadnoughts and vehicles.

I have also tagged Assassins, Knights & Titans as Non-Infantry in their root, and Thallax that a Forgelord can take as Infantry in root.
There might be an issue with Fortifications currently that they can't be tagged either way. But either the category can be moved to GST. Or maybe this is the point in which a "Fortifications Library" could be made, then linked to LA and the other libraries. Then the roots added to each rather than imported.