Berserker Assautl
Predators and Rampages didn't have a root for FA and Troops respectively.

The rampages had a category add for troops and compulsory troops. This didn't work as it was not primary for troops. But also you can still take them in their normal slots, Not just the new slot.
This just required a new root entry.