Needle Pistol, Moriax Knights and Sicaran Venator
Needle Pistol had the wrong range
Moriax Knights incorrected had the Line subtype
It turns out the Sicaran Venator in non-English translations have the Bombard sub-type. Previously in things like the Mech book and the original rulebook in English the foreign language versions have had what turned out to be the correct rules after an FAQ was published. So an option to add it and a rule explanation as to why it is there is included until an FAQ in English or other languages comes out to verify.