More daemonic goodies
All entries for Esoterists, Lorgar Trans and Erebus put in, as well as Rogue Psyker.
Moved a few bits around. Might need to have a play around to get the Aetheric Dominions to play well with Legions. Though shouldn't be to much of an issue with them as it will be more limited. It just means people need to look a bit harder at the rules.