Fix for compatabilty
Fix for #3040 Probably not the best one, and could be cleaned up a bit more as we have some duplicate entries for publications of FAQ and stuff that are X.X and also have an entry of X.X+1. In future the numbers just need updating on the publication, rather than a new entry made to replace. Also the rulebook publication had been deleted (presumably by mistake when things were being tidied up.)