Vigil Storm Shields again
#3153
Issue happened again with the coding of the Termy weapon SGSE for vigil shields in Jan/Feb. This fix was put in to fix that Termy Praetors were mistakenly being given an option to swap their combi weapon for a vigil shield. However in fixing it it then locked out any command squad they took from taking them (though was ok for Artificer Praetors).
It was a problem with "Ancestor Of" being used for the Praetors without making an exception for the command squad that comes below them.
I have made what seems to be a fix by splitting the "OR" for hidden from just Cata or Tart praetor as Ancestor to then have 2 "AND" sections of the same function, but then added 2 additional lines to also check to make sure they are "ancestor is not instance of cata / termy command squad" as well.
I've only tested on BS, and I know NR handles things differently.