Sabre Tanks Pintle Weapons
Fix for Sabre Tanks Pintle Weapons.
There was an issue on BS that when you selected any of the combi weapon options for sabres, it was enforcing the selection on the whole group. This was down to the entries linked were for some reason linked to the "Collective" version of the weapons, and not the individual ones.