Proof of concept for WoD fix for BS users
Had a play around and added a unique version of the heavy stubber to the WoD selection for each unit that seems to work without throwing up errors on BS.
This version I have quickly pieced together that modified it for Tactical, Tact support, Despoilers, Centurions and Praetors.
Though I haven't removed the old heavy stubber option. I have made a unique one for the Tact, Tact support and despoiler squad as a proof of concept.
If this works ok, then I can role out this to all the other 30-40 entries required. Though that is only split over about 10-15 units I think, so wouldn't take long.