Legiones Auxulia
Standard thing of needing line breaks removed.
This adds all the Legiones Auxilia rules as a sub entry to the Solar Auxilia entry.
I had to add 18 new categories to the GST (1 for each legion).
Then had to add to each of the legion roots in their own library to add that category to the legion IF it was also a Crusade detachment in the same was as the ZM tag is added.

This has not added the Overseer. But I have left room for it to be added easily to the unit root entries.