fix for #3313 (Aug Transport)
Augmetic Transport rule non Black shield factions #3313

Just edited all the entries to reverse the logic to fix it.
BS doesn't like a rule that isn't hidden, but the rule link being hidden with a modifier to unhide it.
These should be left as not hidden and the hide modifier added to the entry.