Dark Scions of the Questoris & Automata Malefica
Checked the Dark Scions of the Questoris (All the same as before but updated the publication data.

Automata Malefica
Scions Of The Dark Forges Warlord Traits added,
The Corrupted Engine Unit Sub-type checked for changes Brass Scorpion now has a * about attacks included in profile. updated the publication data
Malefic Aegis (X) (remained the same, but updated the publication data)
The Ætheric Dominion (X) Special Rule updated the publication data but this one might be an odd one as it has left out the whole thing of having a max of 1 type in an army. Though it does acknowledge that it is being brief as there are only 2 of them listed.