Nekrotek, Malamina, Æthertek.
Completed Nekrotek, Malamina, Æthertek

Malamina was the most awkward as it required a rework of the Vorax unit. Though the final result of it is a lot cleaner as you now select which model has an irad if you take them with the new mixed weapons.

Also removed some stuff like the importing of things like the Brass Scorp.
Also had to do a bit of work on the root entry of the Decimator as a new way to select it is available.