Daemons
Changed the way daemons work.
Slightly proof of concept, but needs some work for BS. Might work fine on NR.
Basically removed the "Units" version of the dominions. Added a "Add category" to all the units to add if that dominion is chosen.
This will add the "Heavy" or "Hammer of Wraths" as appropriate.

To Do:
Need to add the dominions to Esoteris, Lorgar, Ebebus.
Need to get the HoW and Heavy to add in Militia & SM forces (not sure why its not working. Might be a bug in BS for Militia)
Strange bug (again might be BS) that the last unit added doesn't get the HoW or Heavy they should get if dominion is picked. Seems that you have to update the unit for it to apply to the last unit, or add an additional unit then delete it. Or reload the rosters.

Keeps brute armaments fix