Rework and updates
Reworked the Infernal Tempest to display more accurately, This is in 2 ways.
Immaterial wings also adds an additional HoW or +1 to any HoW a model already has.
HoW from Infernal Tesmpest also adds an additional HoW or +1 to any HoW a model already has. But in addition also makes all the HoW on the model have Deflagerate.
I have also added a note to indicate that on occasions you will need to add or remove a model in 1 unit to get the army to have the HoW to display correctly. This seems intermittant.

Unforutunately I can't seem to get the "Heavy" subtype added to all models when Putrid Corruption is taken. I am unsure why it works sometimes but now others.
I have added a note to indicate that to the Dominion.